Migrate RouterLink to TypeScript
